requirements.txt
matplotlib
numpy
pyaudio
pyserial

